+++ packages/cater-runtime/index.js
@@ -1,0 +1,12 @@
+// Copyright Jon Williams 2017. See LICENSE file.
+
+const HandlerCater = require('./src/handler-cater');
+const start = require('./src');
+
+const runtime = function() {
+  return start();
+}
+
+runtime.HandlerCater = HandlerCater;
+
+module.exports = runtime;