+++ packages/cater-runtime/src/cater-context.js
@@ -6,7 +6,8 @@
  */
 class CaterContext {
 
-  constructor() {
+  constructor(bundlePath) {
+    this.bundlePath = bundlePath;
     this.globalStyles = [];
   }