+++ packages/cater/server/global-style.js
@@ -1,6 +1,6 @@
 // Copyright Jon Williams 2017. See LICENSE file.
 import React from "react";
-import { caterContextTypes } from './cater-provider'; // Relative import
+import { caterContextTypes } from "server/cater-provider";
 import PropTypes from "prop-types";
 
 /**